% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cah_files}
\alias{cah_files}
\title{Get metadata on CAH files and folders}
\usage{
cah_files(...)
}
\arguments{
\item{...}{The short name(s), quoted or unquoted, of one or more files and/or folders you want metadata for. If nothing specified, all files and folders will be returned.}
}
\value{
A dataframe with metadata on the folders.
}
\description{
Get metadata on CAH files and folders
}
\examples{
# Specify your local paths to the `CAH Shared` folder in Box
# and the `CAH Shared Drive` folder in Drive.
# Best to do this in your .Rprofile file (see README).
# Something like this (but then with your path):
options(
  cah.box_path = "/Users/jwl38/Library/CloudStorage/Box-Box/CAH/CAH Shared",
  cah.drive_path = paste0(
     "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com",
     "/.shortcut-targets-by-id/1mCSu5zzZGk4BAPPBaZCMQXwnKmOPQSfU/CAH/CAH Shared Drive"
  )
)
cah_files()
cah_files(box)
cah_files(projects_sheet)
}
