% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cah_install}
\alias{cah_install}
\title{(Re)install the \code{cah} and \code{better} packages}
\usage{
cah_install(..., build_vignettes = TRUE)
}
\arguments{
\item{...}{Packages you want to (re)install, unquoted. Options: \code{cah}, \code{better}.}

\item{build_vignettes}{Whether or not to build vignettes when installing, by default TRUE, and passed on to \code{remotes::install_github()}.}
}
\value{
Package(s) installed, string, invisible.
}
\description{
Wrapper around \code{remotes::install_github}, which you may know as \code{devtools::install_github}. Since the \code{cah} and \code{better}  packages are currently in development, they may change often, and you may want to update to the latest version often.
}
\examples{
\dontrun{
cah_install()
cah_install(better)
cah_install(cah, better)
}
}
