% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cah_path}
\alias{cah_path}
\title{Get the paths of CAH files or folders}
\usage{
cah_path(...)
}
\arguments{
\item{...}{The short name(s), quoted or unquoted, of one or more files and/or folders you want paths for. If nothing specified, all paths will be returned.}
}
\value{
The path(s) as a string(s).
}
\description{
Get the paths of CAH files or folders
}
\examples{
# In your R scripts you can refer to paths in folders other than the one your script is located in.
# Other people can still run your code.
# Something like this (but then with your paths):
options(
  cah.box_path = "/Users/jwl38/Library/CloudStorage/Box-Box/CAH/CAH Shared",
  cah.drive_path = paste0(
     "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com",
     "/.shortcut-targets-by-id/1mCSu5zzZGk4BAPPBaZCMQXwnKmOPQSfU/CAH/CAH Shared Drive"
  )
)
readr::read_csv(paste0(
  cah_path(box),
  "/IRB\ Projects/Health\ Projects/Health\ Team/Public\ datasets\ on\ health",
  "/Global\ Burden\ of\ Disease\ Study\ Data",
  "/Raw\ Data\ GBD/IHME-GBD_2019_DATA-1a56c841-1/IHME-GBD_2019_DATA-1a56c841-1.csv"
  ))
}
