% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cah_set_options}
\alias{cah_set_options}
\title{Set global options}
\usage{
cah_set_options(force = FALSE)
}
\arguments{
\item{force}{Whether or not to force setting cah options if options are already found}
}
\description{
Sets global options for things like the \code{better} package, specifying the default paths \code{better} will use in many of its functions. Automatically called when loading \code{cah}.
}
\examples{
# Normally the options are already set when loading `cah`.
# But you can also do it manually, perhaps after you changed some paths.
# `cah_set_options` assumes you already specified your local path
# to the `CAH Shared` folder in Box.
# Hopefully you did this in your .Rprofile file (see README).
# Something like this (but then with your path):
options(
  cah.box_path = "/Users/jwl38/Library/CloudStorage/Box-Box/CAH/CAH Shared"
)
# If you did this, you can now set `better` options:
cah_set_options()
}
