% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cah_vignette}
\alias{cah_vignette}
\title{Get pre-installed \code{cah} vignettes}
\usage{
cah_vignette(...)
}
\arguments{
\item{...}{The name(s) of the vignette(s) you want to read, unquoted. By default, if you don't specify anything, it's all the vignettes. Options are: better-demo, cah.}
}
\value{
Names of the vignettes, invisible.
}
\description{
Open the HTML files of pre-installed vignettes.
}
\details{
Some \code{cah} vignettes require things that you may not had set up yet at the time of installing \code{cah}. Since such vignettes can't be built at the time of installation (which is how regular vignettes are created), they are pre-installed as HTML files. Use \code{cah_vignette} to access them. Strictly speaking, they aren't vignettes. You can't access them with \verb{vignette(*, package = "cah")}. But they are articles that fulfill the role of typical vignettes. Note that you can also use \code{cah_vignette()} for true vignettes that are also accessible with \verb{vignette(*, package = "cah")}.
}
\examples{
\dontrun{
cah_vignette()
cah_vignette(cah)
cah_vignette(`better-demo`)
cah_vignette(cah, "better-demo")
}
}
