% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_better}
\alias{install_better}
\title{Install the \code{better} package}
\usage{
install_better(...)
}
\arguments{
\item{...}{Optional arguments passed on to \code{remotes::install_github}.}
}
\value{
Package name returned by \code{remotes::install_github}, invisible.
}
\description{
Wrapper around \code{remotes::install_github}, which you may know as \code{devtools::install_github}. Since the \code{better} package is currently in development, it may change often, and you may want to update to the latest version often.
}
\examples{
\dontrun{
install_better()
}
}
