% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_cah}
\alias{theme_cah}
\title{CAH theme for ggplot}
\usage{
theme_cah()
}
\value{
Plot
}
\description{
CAH theme for ggplot
}
\examples{
library(tidyverse)
p1 <- tibble(
  gender = c(rep("Male",10),rep("Female",10)),
  x = c(1:10,1:10),
  y = c(10:1,1:10)
  ) \%>\%
  ggplot(aes(x,y,color=gender)) +
  geom_line() +
  labs(x = "Some variable", y = "Some other variable", title = "Some title")
p1 # default theme
p1 + theme_cah()
}
