% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cah-package.R
\docType{package}
\name{cah-package}
\alias{cah}
\alias{cah-package}
\title{cah: Useful tools for CAHers}
\description{
The `cah` package contains useful tools for CAHers. The two main categories of tools are: tools that make statistical analysis easier, and tools that make collaboration easier. In addition, there's tools that help with research project coorination, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/janlindemans/cah}
  \item Report bugs at \url{https://github.com/janlindemans/cah/issues}
}

}
\author{
\strong{Maintainer}: Jan Lindemans \email{janwillem.lindemans@gmail.com} (\href{https://orcid.org/0000-0002-3702-8651}{ORCID})

}
\keyword{internal}
